+++ library/src/main/java/org/mustangproject/BankDetails.java
@@ -1,5 +1,7 @@
 package org.mustangproject;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.mustangproject.ZUGFeRD.IZUGFeRDTradeSettlementPayment;
 
 /**
@@ -18,6 +20,10 @@ public class BankDetails implements IZUGFeRDTradeSettlementPayment {
 	 * the "name" of the bank account (holder)
 	 */
 	protected String accountName=null;
+	/***
+	 * bean constructor
+	 */
+	public BankDetails() {	}
 
 	/***
 	 * constructor for IBAN only :-)
@@ -80,12 +86,14 @@ public class BankDetails implements IZUGFeRDTradeSettlementPayment {
 	 * */
 	@Override
 	@Deprecated
+	@JsonIgnore
 	public String getOwnBIC() {
 		return getBIC();
 	}
 
 	@Override
 	@Deprecated
+	@JsonIgnore
 	public String getOwnIBAN() {
 		return getIBAN();
 	}