+++ History.md
@@ -1,14 +1,28 @@
-- #420
-- #441
+- 461
+- 463
+
+2.13.0
+=======
+2024-08-28
+
+- Item Attributes and Country of Origin missing on Product.  #420
+- Avoid NullPointerException if dueDate is not set.  #441
+- support reasoncodes #431 
 - Enhance Charges/Allowances with reasonCode. #432
 - Fix build warnings from editing and building. #415
 - ZUGFeRDVisualizer.toPDF(): generate PDF/A-3b. #400
 - allow access to invoice attachments via  ZUGFeRDInvoiceImporter zii.getFileAttachmentsPDF() 
   and XML (zii.getFileAttachmentsXML)
--  #436 and #370. (langfr)
+- No interface for required field CreditorReferenceID #436 and
+- X-Rechnung direct-debit missing mandatory field BT-90  #370. (langfr)
 - refactor(ZUGFeRDVisualizer): improve PDF visualization performance #438
-
-
+- product creation without description now possible empty description
+- filename of embedded file was not xrechnung.xml when using profile xrechnung #452
+- allow legalorganisation to have a tradingbusinessname #447 
+- JSon deserialization does not work with BankDetails #455
+- Fix ClassCastException in CLI (Main.java). #451
+- changed additional references by line from String to List and implemented it on Item #454
+ 
 2.12.0
 =======
 2024-07-20